refactoring. remove unused code
