+++ internal/app/upgradeagent/container.go
@@ -1,19 +1,14 @@
-package core
+package upgradeagent
 
 import (
 	"fmt"
-	"strconv"
 	"strings"
 
 	"github.com/docker/docker/api/types"
-	dockercontainer "github.com/docker/docker/api/types/container"
 )
 
 const (
-	patraoAgentLabel = "com.centurylinklabs.patrao_agent"
-	signalLabel      = "com.centurylinklabs.watchtower.stop-signal"
-	enableLabel      = "com.centurylinklabs.watchtower.enable"
-	zodiacLabel      = "com.centurylinklabs.zodiac.original-image"
+	signalLabel = "com.centurylinklabs.watchtower.stop-signal"
 )
 
 // NewContainer returns a new Container instance instantiated with the
@@ -43,66 +38,17 @@ func (c Container) Name() string {
 	return c.containerInfo.Name
 }
 
-// ImageID returns the ID of the Docker image that was used to start the
-// container.
-func (c Container) ImageID() string {
-	return c.imageInfo.ID
-}
-
 // ImageName returns the name of the Docker image that was used to start the
 // container. If the original image was specified without a particular tag, the
 // "latest" tag is assumed.
 func (c Container) ImageName() string {
-	// Compatibility w/ Zodiac deployments
-	imageName, ok := c.containerInfo.Config.Labels[zodiacLabel]
-	if !ok {
-		imageName = c.containerInfo.Config.Image
-	}
+	imageName := c.containerInfo.Config.Image
 
 	if !strings.Contains(imageName, ":") {
 		imageName = fmt.Sprintf("%s:latest", imageName)
 	}
 
 	return imageName
-}
-
-// Enabled returns the value of the container enabled label and if the label
-// was set.
-func (c Container) Enabled() (bool, bool) {
-	rawBool, ok := c.containerInfo.Config.Labels[enableLabel]
-	if !ok {
-		return false, false
-	}
-
-	parsedBool, err := strconv.ParseBool(rawBool)
-	if err != nil {
-		return false, false
-	}
-
-	return parsedBool, true
-}
-
-// Links returns a list containing the names of all the containers to which
-// this container is linked.
-func (c Container) Links() []string {
-	var links []string
-
-	if (c.containerInfo != nil) && (c.containerInfo.HostConfig != nil) {
-		for _, link := range c.containerInfo.HostConfig.Links {
-			name := strings.Split(link, ":")[0]
-			links = append(links, name)
-		}
-	}
-
-	return links
-}
-
-// IsPatraoUpgradeAgent returns a boolean flag indicating whether or not the current
-// container is the watchtower container itself. The watchtower container is
-// identified by the presence of the "com.centurylinklabs.watchtower" label in
-// the container metadata.
-func (c Container) IsPatraoUpgradeAgent() bool {
-	return ContainsPatraoAgentLabel(c.containerInfo.Config.Labels)
 }
 
 // StopSignal returns the custom stop signal (if any) that is encoded in the
@@ -114,77 +60,4 @@ func (c Container) StopSignal() string {
 	}
 
 	return ""
-}
-
-// Ideally, we'd just be able to take the ContainerConfig from the old container
-// and use it as the starting point for creating the new container; however,
-// the ContainerConfig that comes back from the Inspect call merges the default
-// configuration (the stuff specified in the metadata for the image itself)
-// with the overridden configuration (the stuff that you might specify as part
-// of the "docker run"). In order to avoid unintentionally overriding the
-// defaults in the new image we need to separate the override options from the
-// default options. To do this we have to compare the ContainerConfig for the
-// running container with the ContainerConfig from the image that container was
-// started from. This function returns a ContainerConfig which contains just
-// the options overridden at runtime.
-func (c Container) runtimeConfig() *dockercontainer.Config {
-	config := c.containerInfo.Config
-	imageConfig := c.imageInfo.Config
-
-	if config.WorkingDir == imageConfig.WorkingDir {
-		config.WorkingDir = ""
-	}
-
-	if config.User == imageConfig.User {
-		config.User = ""
-	}
-
-	//if sliceEqual(config.Cmd, imageConfig.Cmd) {
-	//	config.Cmd = nil
-	//}
-
-	//if sliceEqual(config.Entrypoint, imageConfig.Entrypoint) {
-	//		config.Entrypoint = nil
-	//	}
-
-	//	config.Env = sliceSubtract(config.Env, imageConfig.Env)
-
-	//	config.Labels = stringMapSubtract(config.Labels, imageConfig.Labels)
-
-	//	config.Volumes = structMapSubtract(config.Volumes, imageConfig.Volumes)
-
-	// subtract ports exposed in image from container
-	for k := range config.ExposedPorts {
-		if _, ok := imageConfig.ExposedPorts[k]; ok {
-			delete(config.ExposedPorts, k)
-		}
-	}
-	for p := range c.containerInfo.HostConfig.PortBindings {
-		config.ExposedPorts[p] = struct{}{}
-	}
-
-	config.Image = c.ImageName()
-	return config
-}
-
-// Any links in the HostConfig need to be re-written before they can be
-// re-submitted to the Docker create API.
-func (c Container) hostConfig() *dockercontainer.HostConfig {
-	hostConfig := c.containerInfo.HostConfig
-
-	for i, link := range hostConfig.Links {
-		name := link[0:strings.Index(link, ":")]
-		alias := link[strings.LastIndex(link, "/"):]
-
-		hostConfig.Links[i] = fmt.Sprintf("%s:%s", name, alias)
-	}
-
-	return hostConfig
-}
-
-// ContainsPatraoAgentLabel takes a map of labels and values and tells
-// the consumer whether it contains a valid watchtower instance label
-func ContainsPatraoAgentLabel(labels map[string]string) bool {
-	val, ok := labels[patraoAgentLabel]
-	return ok && val == "true"
 }
