+++ internal/app/upgradeagent/objects.go
@@ -1,4 +1,4 @@
-package core
+package upgradeagent
 
 import (
 	"fmt"
@@ -61,24 +61,6 @@ type ContainerSpec struct {
 	Image string
 	//TBD
 }
-
-/*
-// CurrentContainerVersion structure present container infor for request to upstream api
-type CurrentContainerVersion struct {
-	ID      string
-	CREATED string
-	NAME    string
-	IMAGE   string
-}
-
-// NewContainerVersion structure present information about new version of running container
-type NewContainerVersion struct {
-	ID            string
-	NAME          string
-	IMAGE         string
-	DeleteVolumes bool `json:"DELETE_VOLUMES"`
-}
-*/
 
 // UpstreamResponseUpgradeInfo structure represent response from Upstream Service
 type UpstreamResponseUpgradeInfo struct {
