+++ internal/app/upgradeagent/upgradeagent.go
@@ -1,4 +1,4 @@
-package core
+package upgradeagent
 
 import (
 	"encoding/json"
@@ -6,7 +6,6 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"time"
 
@@ -42,7 +41,11 @@ func runOnce(context *cli.Context) error {
 		log.Error(rc)
 		return rc
 	}
-	upgradeInfoArray, rc := getLaunchedSolutionsList(&containers, context)
+	if len(containers) == 0 {
+		log.Info("There are no launched containers on the host")
+		return nil
+	}
+	upgradeInfoArray, rc := getLaunchedSolutionsList(context, &containers)
 	if nil != rc {
 		log.Error(rc)
 		return rc
@@ -54,8 +57,10 @@ func runOnce(context *cli.Context) error {
 
 func schedulePeriodicUpgrades(context *cli.Context) error {
 	log.Infoln("[+]schedulePeriodicUpgrades()")
-	gocron.Every(uint64(context.GlobalInt64(UpgradeIntervalName))).Seconds().Do(runOnce, context)
-	<-gocron.Start()
+	{
+		gocron.Every(uint64(context.GlobalInt64(UpgradeIntervalName))).Seconds().Do(runOnce, context)
+		<-gocron.Start()
+	}
 	log.Infoln("[-]schedulePeriodicUpgrades()")
 	return nil
 }
@@ -88,31 +93,25 @@ func getSolutionAndServiceName(containerName string) (string, string, error) {
 	return nameParts[0][1:], nameParts[1], err
 }
 
-func getLaunchedSolutionsList(containers *[]Container, context *cli.Context) ([]UpstreamResponseUpgradeInfo, error) {
+func getLaunchedSolutionsList(context *cli.Context, containers *[]Container) ([]UpstreamResponseUpgradeInfo, error) {
 	var (
 		solutions []string
 		rc        []UpstreamResponseUpgradeInfo
 		err       error
 	)
-
-	log.Info("[+]getLaunchedSolutionsList")
 	getURLPath := context.GlobalString(UpstreamName) + UpstreamGetUpgrade
 
 	for _, current := range *containers {
 		currentSolution, _, err := getSolutionAndServiceName(current.Name())
-
 		if nil != err {
 			log.Error(err)
 			continue
 		}
-
 		if contains(&solutions, &currentSolution) {
 			continue
 		}
-
 		solutions = append(solutions, currentSolution)
 		currentPath := getURLPath + currentSolution
-
 		resp, err := http.Get(currentPath)
 		if nil != err {
 			log.Error(err)
@@ -137,14 +136,11 @@ func getLaunchedSolutionsList(containers *[]Container, context *cli.Context) ([]
 
 		rc = append(rc, toUpgrade)
 	}
-	log.Info("[-]getLaunchedSolutionsList")
 	return rc, err
 }
 
 func doUpgradeSolutions(upgradeInfoList *[]UpstreamResponseUpgradeInfo, containers *[]Container) error {
 	var rc error
-
-	log.Info("[+]doUpgradeSolutions")
 	for _, item := range *upgradeInfoList {
 		if !isNewVersion(&item, containers) {
 			log.Infof("Solution [%s] is up-to-date.", item.Name)
@@ -153,57 +149,19 @@ func doUpgradeSolutions(upgradeInfoList *[]UpstreamResponseUpgradeInfo, containe
 		for _, container := range *containers {
 			name, _, _ := getSolutionAndServiceName(container.Name())
 			if item.Name == name {
-				err := client.StopContainer(container, 0)
+				err := client.StopContainer(container, StopContainerDefaultTimeoutS)
 				if nil != err {
 					log.Error(err)
-					// TBD
-					continue
 				}
 			}
 		}
-		if _, isFileExist := os.Stat(item.Name); !os.IsNotExist(isFileExist) {
-			os.RemoveAll(item.Name)
-		}
-		err := os.Mkdir(item.Name, os.ModePerm)
-		if nil != err {
-			log.Error(err)
-			// TBD
-			continue
-		}
-		defer os.Remove(item.Name)
-		dockerComposeFileName := fmt.Sprintf("%s/%s", item.Name, DockerComposeFileName)
-
-		f, err := os.Create(dockerComposeFileName)
+		err := client.LaunchSolution(&item)
 		if err != nil {
 			log.Error(err)
-			continue
-			//TBD
-		}
-		defer func() {
-			f.Close()
-			os.Remove(dockerComposeFileName)
-		}()
-
-		_, err = f.Write([]byte(item.Spec)[:len(item.Spec)])
-		if nil != err {
-			log.Error(err)
-			// TBD
-			continue
-		}
-
-		log.Infof("Launching solution [%s]", item.Name)
-
-		cmd := exec.Command(DockerComposeCommand, "-f", fmt.Sprintf("%s/%s", rootPath, dockerComposeFileName), "up", "-d")
-		err = cmd.Run()
-
-		if nil != err {
-			log.Error(err)
-			// TBD
 			continue
 		}
 		log.Infof("Solution [%s] is successful launched", item.Name)
 	}
-	log.Info("[-]doUpgradeSolutions")
 	return rc
 }
